(PC-23190)[API] chore: extract get product by ean test
